Enforce eval shuffle warning only for default samplers (#12653)
